+++ sdk/formrecognizer/ai-form-recognizer/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.1.0-beta.1 (Unreleased)
 
+- Added an `appearance` property to `FormLine` that contains information about the appearance of the line, such as style (e.g. "handwritten").
 - Added a `boundingBox` property to `FormTable` that has a bounding box that contains the entire table.
 - Added support for the "image/bmp" content type. This content type is supported on all methods that accept a `FormRecognizerRequestBody` **except** for custom form recognition.
 - Added a `language` option to `BeginRecognizeContentOptions`. By default, when performing layout/content analysis, the service will attempt to detect the language of the document and supports multi-language inputs. The `language` parameter allows you to override this behavior and force the service to use a specific language.